original-inventory/backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.40.0
anthropic>=0.30.0
python-dotenv>=1.0.0
